+++ lib/mssql.dart
@@ -8,3 +8,5 @@ export 'src/connection.dart';
 export 'src/exception.dart';
 export 'src/pool.dart';
 export 'src/result.dart';
+export 'src/server_endpoint.dart';
+export 'src/tds/sql_browser.dart';